Change load to be middleware based
