+++ components/settings.jsx
@@ -1,6 +1,5 @@
 import React, { Component, PropTypes } from 'react'
-import loadArchive from '../middleware/loadArchive'
-import { setArchiveDisplayPath } from '../actions/index'
+import { setArchiveDisplayPath, loadArchive } from '../actions/index'
 
 const divStyle = {padding:10}
 const inputStyle = {width:"100%"}
@@ -20,8 +19,8 @@ export default class Settings extends Component {
 					style={inputStyle}
 					value={archive.displayPath}
 					onChange={(event)=>{store.dispatch(setArchiveDisplayPath(event.target.value))}}
-				 	onKeyPress={(event)=>{ if (event.which === 13) loadArchive(store, event.target.value) }}
-					onBlur={(event)=>{loadArchive(store, event.target.value)}}
+				 	onKeyPress={(event)=>{ if (event.which === 13) store.dispatch(loadArchive(event.target.value)) }}
+					onBlur={(event)=>{ store.dispatch(loadArchive(event.target.value)) }}
 				/>
 				<p hidden={settings.loadArchiveErrorMsg === undefined} style={errorStyle}>
 					{settings.loadArchiveErrorMsg}
