+++ middleware/getAllMiddleware.js
@@ -1,0 +1,4 @@
+import { applyMiddleware } from 'redux';
+import loadArchiveMiddleware from './loadArchiveMiddleware'
+
+export default () => applyMiddleware(loadArchiveMiddleware)
